refactor: 不要ファイル削除 プリティア レイアウト調整
