+++ src/components/AuthButton.tsx
@@ -24,11 +24,13 @@ export const AuthButton = () => {
     );
   }
   return (
-    <button
-      className="px-4 py-2 bg-blue-500 text-white rounded-lg hover:bg-blue-600"
-      onClick={() => signIn('auth0', { redirect: false })}
-    >
-      サインイン
-    </button>
+    <div className="flex justify-end">
+      <button
+        className="px-4 py-2 bg-blue-500 text-white rounded-lg hover:bg-blue-600"
+        onClick={() => signIn('auth0', { redirect: false })}
+      >
+        サインイン
+      </button>
+    </div>
   );
 };
