+++ src/pages/setting.tsx
@@ -1,6 +1,6 @@
 import { useState } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
-import { selectUser, updateUsername } from '../../../store/userSlice';
+import { selectUser, updateUsername } from '../store/userSlice';
 import { useSession } from 'next-auth/react';
 import { useRouter } from 'next/router';
