+++ src/styles/globals.css
@@ -7,6 +7,6 @@ body {
   @apply bg-gray-100 text-gray-900 transition-colors duration-300 ease-in-out;
 }
 
-.dark body {
+.dark body, textarea {
   @apply bg-gray-900 text-gray-100; /* ダークモード */
 }
